chore: OrderPage 제거 및 RouletteResultPage에 메인화면 리디렉션 버튼 추가
